(Onboarding): split stepper view component
